added a template for associatedProject and included it on the project summary tab.
The basic lter project description had an open query (?) and so if there was an associatedProject, that project's content displayed on the same page as the chosen project.

I constructed the url for the related project with variables, and this is not a generic solution. So I will not add this to the lterProjectDescription stlyesheet at this time.